feat: add reusable exposed drop down menu
